refactor: Subfolders for Environment + adjusted namespaces.
