--- src/Syringe.Core/Environment/IOctopusRepositoryFactory.cs
@@ -1,9 +1,0 @@
-using Octopus.Client;
-
-namespace Syringe.Core.Environment
-{
-    public interface IOctopusRepositoryFactory
-    {
-        IOctopusRepository Create();
-    }
-}
