+++ src/Syringe.Core/Environment/Octopus/IOctopusRepositoryFactory.cs
@@ -1,0 +1,9 @@
+using Octopus.Client;
+
+namespace Syringe.Core.Environment.Octopus
+{
+    public interface IOctopusRepositoryFactory
+    {
+        IOctopusRepository Create();
+    }
+}
