+++ src/Syringe.Service/DependencyResolution/DefaultRegistry.cs
@@ -21,6 +21,8 @@ using StructureMap;
 using StructureMap.Graph;
 using Syringe.Core.Configuration;
 using Syringe.Core.Environment;
+using Syringe.Core.Environment.Json;
+using Syringe.Core.Environment.Octopus;
 using Syringe.Core.IO;
 using Syringe.Core.Runner;
 using Syringe.Core.Runner.Logging;
