+++ src/Syringe.Tests/Unit/Service/DependencyResolution/DefaultRegistryTests.cs
@@ -5,6 +5,7 @@ using Octopus.Client;
 using StructureMap;
 using Syringe.Core.Configuration;
 using Syringe.Core.Environment;
+using Syringe.Core.Environment.Octopus;
 using Syringe.Core.IO;
 using Syringe.Core.Runner.Logging;
 using Syringe.Core.Tests.Repositories;
